requirements.txt
pymongo==3.5.1
prompt_toolkit==3.0.24
Pillow==9.0.1